requirements.txt
Flask
Flask-cors
python-dotenv
langchain_community
sentence_transformers
groq
tiktoken
pyodbc
markdown2
PyPDF2==3.0.1
flask-jwt-extended==4.6.0
qdrant-client
openai
